pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "reels-gen"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "moviepy>=2.0.0",
    "Pillow>=10.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
reels-gen = "reels_gen.cli:main"

[tool.setuptools.packages.find]
include = ["reels_gen*"]
